\name{plot.bgmycprobmat}
\alias{plot.bgmycprobmat}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
plots the matrix of probability of conspecificity}
\description{
plots the matrix of probability of conspecificity along with a tree and a legend in order to visualize results}
\usage{
plot.bgmycprobmat(probmat, tree)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{probmat}{
output from spec.probmat}
  \item{tree}{
a single tree in "phylo" object format}
}
\details{
a plotting function that visualizes the matrix of probabilities of conspecificity.}
\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
Noah M. Reid}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{

}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
