\name{plot.bgmycrates}
\alias{plot.bgmycrates}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
plots relative coalescent and Yule rates for model fit evaluation}
\description{
plots relative coalescent and Yule rates for model fit evaluation}
\usage{
plot.bgmycrates(checkratesoutput)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{checkratesoutput}{
output from checkrates function}
}
\details{
plots relative coalescent and Yule rates for model fit evaluation}
\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
Noah M. Reid}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{

}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
